Add better threading model and bucketing
This commit changes the threading model so that instead of assigning a
list of pairs to process to each thread, we instead have one queue
(a vector) for all pairs, and have each thread fetch pairs from that
when they are done with processing the last one. We also no longer
recurse a pair all the way down the tree of child pairs, but instead
each thread adds all child pairs to the work queue so they can be
processed separately.

This commit also adds the option to split the work into "buckets" where
the list of initial pairs is split into a number of units (buckets)
that can be processed separately. The new CLI handles can be used to
define the number of buckets to split the work into, and which bucket to
process.

The bucketing is done so that the initial pairs are spread around the
buckets to prevent all sibling pairs from ending up in the same bucket.

This, in theory, should spread the amount of work needed roughly evenly
to each bucket.