Add make commands to download dependencies (fmtlib/fmt and boost)
This commit also modies the compilation options and main.cpp includes to
use the downloaded libs instead of expecting global ones.

The GMP libs are still expected to be available globally, mostly because
they likely won't be used and I can't be arsed to add the as local too.

This is needed to make compiling this code on the CSC machines smooth.