Add subdivide_with_cutoff_condition function
The new function differs from the subdivide function by accepting a
function as an argument that, if it returns true, causes the subdivision
to stop.

This will be used to limit the number of child pairs needed as they can
in some cases be known to meet the Littlewood criteria. This decreases
the needed computation by 15% or so.