Return more info from the Littlewood criteria check
This commit changes the return type of the meets_littlewood_criteria
function from bool to a struct that also contains the "best Q" value.

The best Q value is the value for Q that results in the lowest value for
the littlewood() function.

This value can be used to limit the subdivision to child pairs.