Change initial convergent pair formulation
After this change we discard the convergent pairs that are below the
identity line, as that will leave out pairs that are symmetric with
pairs above the identity line.

Doing this require that we have a fine enough division of the initial
convergents, so we first need to "refine" the convergents until their
denominators are large enough.