Remove subdivide() method and rename subdivide_with_cutoff_condition()
Since the check_pair method removed in the previous commit was the only
remaining part of the code using the subdivide() method, it was also
unnecessary.

Since that name was now freed, we also renamed the
subdivide_with_cutoff_condition() to just subdivide() (since there is
now no other way to do the subdivision).